Successfully split up App.tsx with updated main
