+++ Isle of Trust/src/components/sideBarState.tsx
@@ -1,4 +1,4 @@
-import "./css/App.css";
+import "../css/App.css";
 import {
     Agent,
     Relation,
